extra methods for inference and tracing
