+++ catalyst/dl/runner/supervised.py
@@ -1,19 +1,9 @@
-from typing import Any, Callable, Dict, List, Mapping, Tuple, Union
-from collections import OrderedDict
+from typing import Any, Callable, List, Mapping, Tuple, Union
 import logging
-from pathlib import Path
 
 import torch
-from torch.jit import ScriptModule
-from torch.utils.data import DataLoader
 
-from catalyst.dl import (
-    CheckpointCallback,
-    InferCallback,
-    State,
-    SupervisedExperiment,
-    utils,
-)
+from catalyst.dl import State, SupervisedExperiment
 from catalyst.utils.tools.typing import Device, Model
 
 from .core import Runner
@@ -125,7 +115,7 @@ class SupervisedRunner(Runner):
     def _handle_batch(self, batch: Mapping[str, Any]) -> None:
         """
         Inner method to handle specified data batch.
-        Used to make a train/valid/infer step during Experiment run.
+        Used to make a train/valid/infer stage during Experiment run.
 
         Args:
             batch (Mapping[str, Any]): dictionary with data batches
@@ -155,142 +145,6 @@ class SupervisedRunner(Runner):
         batch = self._batch2device(batch, self.device)
         output = self.forward(batch, **kwargs)
         return output
-
-    def predict_loader(
-        self,
-        model: Model,
-        loader: DataLoader,
-        resume: str = None,
-        verbose: bool = False,
-        state_kwargs: Dict = None,
-        fp16: Union[Dict, bool] = None,
-        check: bool = False,
-    ) -> Any:
-        """
-        Makes a prediction on the whole loader with the specified model.
-
-        Args:
-            model (Model): model to infer
-            loader (DataLoader): dictionary containing only one
-                ``torch.utils.data.DataLoader`` for inference
-            resume (str): path to checkpoint for model
-            verbose (bool): ff true, it displays the status of the inference
-                to the console.
-            state_kwargs (dict): additional state params to ``State``
-            fp16 (Union[Dict, bool]): If not None, then sets inference to FP16.
-                See https://nvidia.github.io/apex/amp.html#properties
-                if fp16=True, params by default will be ``{"opt_level": "O1"}``
-            check (bool): if True, then only checks that pipeline is working
-                (3 epochs only)
-        """
-        loaders = OrderedDict([("infer", loader)])
-
-        callbacks = OrderedDict([("inference", InferCallback())])
-        if resume is not None:
-            callbacks["loader"] = CheckpointCallback(resume=resume)
-
-        self.infer(
-            model=model,
-            loaders=loaders,
-            callbacks=callbacks,
-            verbose=verbose,
-            state_kwargs=state_kwargs,
-            fp16=fp16,
-            check=check,
-        )
-
-        output = callbacks["inference"].predictions
-        if isinstance(self.output_key, str):
-            output = output[self.output_key]
-
-        return output
-
-    def trace(
-        self,
-        model: Model = None,
-        batch=None,
-        logdir: str = None,
-        loader: DataLoader = None,
-        method_name: str = "forward",
-        mode: str = "eval",
-        requires_grad: bool = False,
-        fp16: Union[Dict, bool] = None,
-        device: Device = "cpu",
-        predict_params: dict = None,
-    ) -> ScriptModule:
-        """
-        Traces model using Torch Jit.
-
-        Args:
-            model (Model): model to trace
-            batch: batch to forward through the model to trace
-            logdir (str, optional): If specified,
-                the result will be written to the directory
-            loader (DataLoader, optional): if batch is not specified, the batch
-                will be ``next(iter(loader))``
-            method_name (str): model's method name that will be traced
-            mode (str): ``train`` or ``eval``
-            requires_grad (bool): flag to trace with gradients
-            fp16 (Union[Dict, bool]): If not None, then sets
-                tracing params to FP16
-            device (Device): Torch deivice or a string
-            predict_params (dict): additional parameters for model forward
-        """
-        if batch is None:
-            if loader is None:
-                raise ValueError(
-                    "If batch is not provided the loader must be specified"
-                )
-            batch = next(iter(loader))
-
-        if model is not None:
-            self.model = model
-
-        if isinstance(fp16, bool) and fp16:
-            opt_level = "O1"
-        elif isinstance(fp16, bool) and not fp16:
-            opt_level = None
-        elif isinstance(fp16, dict):
-            opt_level = fp16["opt_level"]
-        else:
-            opt_level = fp16
-
-        if opt_level is not None:
-            device = "cuda"
-        elif device is None:
-            if self.device is None:
-                self.device = utils.get_device()
-            device = self.device
-
-        result = utils.trace_model(
-            model=self.model,
-            runner=self,
-            batch=batch,
-            method_name=method_name,
-            mode=mode,
-            requires_grad=requires_grad,
-            opt_level=opt_level,
-            device=device,
-            predict_params=predict_params,
-        )
-
-        if logdir is not None:
-            filename = utils.get_trace_name(
-                method_name=method_name,
-                mode=mode,
-                requires_grad=requires_grad,
-                opt_level=opt_level,
-            )
-
-            logdir = Path(logdir)
-            output: Path = logdir / "trace"
-            output.mkdir(exist_ok=True, parents=True)
-
-            out_model = str(output / filename)
-
-            torch.jit.save(result, out_model)
-
-        return result
 
 
 __all__ = ["SupervisedRunner"]
